widen the brush check count to match the counter it is compared against
It was an unsigned short against the clip map's int, so once the map counter
passed 65535 the comparison could never hold and every brush was re-tested in
every leaf it spans for the rest of the session.

+++ src/code/qcommon/cm_local.h
@@ -67,7 +67,7 @@ typedef struct cbrush_s {
 	vec3_t				bounds[2];
 	cbrushside_t		*sides;
 	unsigned short		numsides;
-	unsigned short		checkcount;		// to avoid repeated testings
+	int					checkcount;		// matches clipMap_t's, which counts past 65535
 } cbrush_t;
 
 class CCMShader
